+++ code/code.l
@@ -12,11 +12,24 @@ open \<
 compname (Heading|Para|Cards)
 propname (title|content|titleSize|contentSize|titleColor|contentColor)
 close \/\> 
-
+container container 
+spacing spacing=\{[1-9]\}
+direction direction=("row"|"column")
+justify justify=("space-evenly"|"space-between"|"flex-start")
+align alignItems=("center"|"flex-start") 
+cgrid \>
+ctgrid \<\/Grid\>
+grid Grid
 %%
 
 {open} {yylval.opentag = strdup(yytext); return (OPENTAG);}
 {close} {yylval.closetag = strdup(yytext);return (CLOSINGTAG);}
+{grid} {return GRID;}
+{container} {return (CONTAINER);}
+{spacing} {return SPACING;}
+{direction} {return DIRECTION;}
+{justify} {return JUSTIFY;}
+{align} {return ALIGN;}
 {compname} {yylval.comp = strdup(yytext); return COMPNAME;}
 {propname} {yylval.name = strdup(yytext); return (PROPNAME);}
 "=" {return (EQ);}
